requirements.txt
asgiref==3.3.1
backcall==0.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
geographiclib==1.50
geopy==2.1.0
gunicorn==20.1.0
haversine==2.3.0
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
mysqlclient==2.0.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.0
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
rope==0.18.0
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
